requirements.txt
blinker==1.6.2
click==8.1.6
colorama==0.4.6
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==2.3.6
mysql-connector-python==8.2.0
SQLAlchemy==2.0.23
PyMySQL==1.1.0
flask_sqlalchemy
